backend/requirements.txt
# Backend Dependencies for HexaCiphers
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
celery==5.3.1
redis==5.0.0
tweepy==4.14.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
langdetect==1.0.9
requests==2.31.0
aiohttp==3.8.5
gunicorn==21.2.0